Move handle-messages function into its own file.
